ml-service/requirements.txt
Flask==2.3.2
scikit-learn==1.2.2
pandas==2.0.2
numpy==1.24.3
joblib==1.2.0
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.31.0
tensorflow==2.12.0
pytest==7.3.1
pytest-cov==4.1.0
flake8==6.0.0
black==23.3.0
